add basic tests for blueprint builder
